# openspec/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GraphViewer API
  description: API specification for GraphViewer diagram rendering service
  version: 1.0.0
  contact:
    name: GraphViewer Team
    url: https://github.com/LessUp/graph-viewer

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://graph-viewer.example.com
    description: Production server

paths:
  /api/render:
    post:
      summary: Render a diagram via Kroki proxy
      description: |
        Renders diagram code using specified engine and returns the output.
        Uses in-memory cache (TTL 120s) for repeated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - engine
                - format
                - code
              properties:
                engine:
                  type: string
                  description: Diagram engine identifier
                  enum:
                    - mermaid
                    - plantuml
                    - graphviz
                    - d2
                    - nomnoml
                    - ditaa
                    - blockdiag
                    - svgbob
                    - seqdiag
                    - actdiag
                    - nwdiag
                    - wavedrom
                    - erd
                    - vega
                    - vegalite
                    - flowchart
                  example: mermaid
                format:
                  type: string
                  description: Output format
                  enum: [svg, png, pdf]
                  example: svg
                code:
                  type: string
                  description: Diagram source code (max 100KB)
                  maxLength: 102400
                  example: |
                    graph TD
                      A[Start] --> B{Decision}
                      B -->|Yes| C[Action]
                      B -->|No| D[End]
            examples:
              mermaid_flowchart:
                value:
                  engine: mermaid
                  format: svg
                  code: |
                    graph TD
                      A[Start] --> B{Decision}
                      B -->|Yes| C[Action]
              plantuml_sequence:
                value:
                  engine: plantuml
                  format: svg
                  code: |
                    @startuml
                    Alice -> Bob: Hello
                    @enduml
      responses:
        '200':
          description: Successful rendering
          content:
            application/json:
              schema:
                type: object
                required:
                  - contentType
                properties:
                  svg:
                    type: string
                    description: SVG content (when format=svg)
                  base64:
                    type: string
                    description: Base64 encoded content (when format=png/pdf)
                  contentType:
                    type: string
                    example: image/svg+xml
              examples:
                svg_response:
                  value:
                    svg: <svg>...</svg>
                    contentType: image/svg+xml
                png_response:
                  value:
                    base64: iVBORw0KGgo...
                    contentType: image/png
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
              examples:
                engine_not_supported:
                  value:
                    error: "Engine 'unknown' is not supported"
                    code: ENGINE_NOT_SUPPORTED
                code_too_large:
                  value:
                    error: "Code exceeds maximum size of 100KB"
                    code: CODE_TOO_LARGE
        '500':
          description: Rendering failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
              examples:
                render_failed:
                  value:
                    error: "Failed to render diagram"
                    code: RENDER_FAILED
                render_timeout:
                  value:
                    error: "Rendering timed out after 10s"
                    code: RENDER_TIMEOUT
      x-codeSamples:
        - lang: JavaScript
          source: |
            fetch('/api/render', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({
                engine: 'mermaid',
                format: 'svg',
                code: 'graph TD; A-->B'
              })
            })

  /api/healthz:
    get:
      summary: Health check endpoint
      description: Returns health status for Docker/Netlify deployment monitoring
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
              example:
                status: ok